Organizes file structure
Refactors the project structure by introducing a dedicated 'readers' directory.

Moves file reading functionalities into the new directory for better code organization and maintainability.

Improves modularity and separation of concerns within the codebase.
